library_management/requirements.txt
click==8.0.3
psycopg2-binary==2.9.1
